Add t/10-headers.t to MANIFEST, test suite. t/10-headers.t exercises the PHP<->Perl communication through the PHP/header <-> PHP::assign_global channels, which is quite a powerful concept.
Add a header callback registration to t/MojoTestServer.pm
for use with t/10-headers.t

Implement header callbacks in MojoX::Template::PHP. You register
a regular expression and a callback (code reference). When a header
arrives that matches the regular expression, the callback is
invoked with the header key and the header value. The callback
should return false if it is not necessary for MojoX::Template::PHP
to include the header with the response.