Refactor header callback in ::Template::PHP to implement additional callbacks configured in the plugin configuration like
    plugin 'MojoX::Plugin::PHP' => {
        php_var_preprocessor => sub { ... },
	php_stderr_processor => sub { ... },
	php_header_processor => sub { ... },
	php_output_postprocessor => sub { ... }
    };

Callbacks are invoked in  MojoX::Template::PHP  during
the  interpret  method.

MojoX::Plugin::PHP receives this configuration in its  register
method, and saves it in the config of  MojoX::Template::PHP.

Added CONFIG pod section to MojoX::Plugin::PHP describing
these callbacks and what they are good for.

Updates to  t/10-headers.t  and  t/11-globals.t  to conform
to this new scheme of using callbacks.