Add t/23-include.t to test suite and MANIFEST. Include helper files t/templates/include.php, include1.php, include2/include3.php. These are our first test cases NOT copied from Catalyst-View-Template-PHP test suite. In these tests, we test where a PHP template may include other PHP templates without trouble.
Wrote information about the current solution into the PLAN.

Remove deadwood hooks from MojoX::Plugin::PHP.
Refactor the way we remove the *template_XXXXXXX parameter
from the parameter list.

::Plugin::PHP::_template_path method, save the directory
where the template was found. In the  render  method,
we'll change to that directory right before starting
the PHP engine. We'll also include this path in the PHP
include_path.

In the ::Plugin::PHP::_interpret method, continue to
prepend  ?>  to the code to be evaluated, but don't
append  <?php  . It is not strictly necessary, and in
files that don't close their final  <?php  tag (and a lot
of WordPress files are like that), it will be an error.

If the PHP evaluation fails in _interpret, set the
response code to 500.